tools/repo_stats/requirements.txt
typer==0.13.1
python-dotenv==1.0.1
markdown2==2.5.1
google-cloud-storage==2.18.2
pandas==2.2.3
streamlit==1.40.1
